+++ llama_test/test_llama_index_chatbot.py
@@ -12,27 +12,6 @@ from deepeval.tracing import Tracer, GenericAttributes
 from deepeval.integrations.llama_index import LlamaIndexCallbackHandler
 
 ###########################################################
-# set up integration
-###########################################################
-
-
-# set llama index global handler
-def deepeval_callback_handler(**eval_params: Any) -> BaseCallbackHandler:
-    return LlamaIndexCallbackHandler(**eval_params)
-
-
-def set_global_handler(eval_mode: str, **eval_params: Any) -> None:
-    """Set global eval handlers."""
-    if eval_mode == "deepeval":
-        handler = deepeval_callback_handler(**eval_params)
-        import llama_index.core
-
-        llama_index.core.global_handler = handler
-
-
-set_global_handler("deepeval")
-
-###########################################################
 # test chatbot
 ###########################################################
